adrian

+++ EyeKnowRight/Models/Employee.cs
@@ -41,12 +41,12 @@ namespace EyeKnowRight.Models
         public byte[] Picture { get; set; }
         public DateTime? BirthDate { get; set; }
         public int? MyProperty { get; set; }
-        public int MaternityLeave { get; set; }
-        public int PaternityLeave { get; set; }
-        public int SickLeave { get; set; }
-        public int BereavementLeave { get; set; }
-        public int MedicalLeave { get; set; }
-        public int PersonalLeave { get; set; }
+        public float MaternityLeave { get; set; }
+        public float PaternityLeave { get; set; }
+        public float SickLeave { get; set; }
+        public float BereavementLeave { get; set; }
+        public float MedicalLeave { get; set; }
+        public float PersonalLeave { get; set; }
         public int NumberOfTries { get; set; }
         public int DaysContract { get; set; }
         public string SSSNumber { get; set; }
